feat(#1322): struct initializers (ADR-014) move to pass 2.1; fixes #1277
Two throws in `CodeGenerator._resolveStructInitializerTypeName`, both reported
as `1:0`, and one of them fired on valid code. `output/` is at 22, from 184.

## One question, asked once

A struct literal has no type of its own. Either it writes one
(`Point { x: 1 }`) or the position it stands in supplies one
(`Point p <- { x: 1 }`), and exactly one must hold: both is a redundant written
type (E0356), neither cannot be resolved (E0357).

Pass 2.1 asks the STRUCTURAL question -- "does an enclosing position supply a
type?" -- and deliberately not "which type". Asking for the name conflates "no
position supplies one" with "a position supplies one and this pass cannot name
it", and those are opposite answers. That is not hypothetical: a field of a
struct declared in an included C header resolves through the C symbols rather
than the C-Next view, and the first draft reported a false E0357 on nine
`tests/interop/anonymous-structs` fixtures. The corpus caught it, which is what
the corpus is for.

## #1277: three valid shapes were rejected, and the cause was in emission

`return { x: 1, y: 2 };` was rejected as "Cannot infer struct type" while the
identical literal assigned to a local first compiled. Probing every position a
struct literal can occupy found three of thirteen broken, all the same cause --
a type was available and codegen did not thread it:

- a `return` at file scope: the return path set an expected type only when the
  return type was an ENUM
- a `return` inside a scope method: `ScopeGenerator` never recorded the return
  type at all
- a `for` header's declaration: it set no expected type, though a block-level
  declaration does

Relocating the throw without fixing these would have preserved a rejection of
valid code behind a fresh error code.

**The scope-method half was found by an invariant this card installed.** A bare
enum member returned from a scope method reached `Internal:` rather than a
diagnostic, because 2.1 said the position established an enum type and codegen
disagreed. `FunctionGenerator` and `ScopeGenerator` had each open-coded the same
four-step function context in the same order and the scope copy was missing
`setCurrentFunctionReturnType` -- one line, and no `return` in any scope method
knew what it returned. Both now call one `enterFunctionContext`/
`exitFunctionContext` pair, so a fifth fact is one edit rather than two that
have to be remembered together.

## The broad change, taken deliberately

A `return` expression is expected to be the declared return type. That is the
fact; restricting it to enums described the consumers rather than the fact, and
is why #1277 existed. Threading it for every return type moves **505 fixtures**,
and the whole diff was verified line by line rather than eyeballed:

| changed lines | what |
| --- | --- |
| 17321 | a literal gained its MISRA C:2012 Rule 7.2 integer suffix |
| 3264 | a suffix moved within an unchanged expression |
| 38 | a `u64` return's literal gained `ULL` |
| 22 | a narrowing return gained its MISRA 10.3 cast (`static_cast` in C++) |
| 12 | a `bool` return gained its explicit `!= 0U` conversion (MISRA 10.1) |

Every one is a conversion the identical literal already received in
`u8 x <- 1;`. The rule did not reach a return statement only because the type
did not. `npm run validate:c` -- cppcheck, MISRA, clang-tidy, flawfinder over
1623 files -- passes on the regenerated corpus.

## Found here, filed not fixed

**#1530**: a struct literal passed as an argument emits by-value C against
ADR-006's pointer parameter, and does not compile, at exit 0. Found by writing
the execution fixture -- the two error fixtures use that exact form as a
negative control and pass, because a `test-error` fixture never compiles its C.
Exit 0 is not evidence when only an execution fixture reaches the compiler.

## A question for the ADR's owner, raised not answered

**ADR-014's own example is code the transpiler rejects.** Line 162 documents
`Point p <- Point { x: 10, y: 20 };` as the way to write it; that is exactly
what E0356 rejects, and has been rejected since before this card (the throw and
its fixture both predate it). This commit preserves the behavior and does not
adjudicate the divergence.

One consequence is new and named here rather than left to be discovered: with a
return now supplying a type, `return Point {};` is redundant too, which removes
the last position where the explicit empty form `Type {}` could be written
against a declared type. A unit test asserted that position, describing it as
"the only way to use explicit type syntax without expectedType context" -- true
only because the return type was not threaded.

Mutation table, one invocation per directory, mutation asserted applied and
restore asserted byte-identical:

| mutation | adr-014 | interop | structs | bugs |
| --- | --- | --- | --- | --- |
| E0356 disabled | 1 | 0 | 0 | 0 |
| E0357 disabled | 1 | 0 | 0 | 0 |
| return no longer establishes a type | 3 | 0 | 0 | 0 |
| field initializer no longer establishes a type | 1 | 0 | 0 | 0 |
| subscript no longer clears | 1 | 0 | 0 | 0 |
| baseline | 0 | 0 | 0 | 0 |

The last two reddened NOTHING on the first run. Both guards were real and the
fixtures simply never built their contexts -- an inferred initializer inside an
explicitly typed one is the only shape where the enclosing field is the sole
establishing node, and a struct literal inside a subscript is the only shape
where the walk must stop rather than climb to the declaration. Both are
fixtures now.

1227/1227 fixtures, 7166 unit tests, oxlint and knip clean.

+++ tests/bugs/issue-1333-scope-reopening/cross-file-span.expected.cpp
@@ -58,7 +58,7 @@ Lib__Point Lib__makeOrigin(void) {
 uint8_t Lib__useFlags(void) {
     Lib__Flags f = 0;
     f = (f & ~(1U << 0)) | (1U << 0);
-    return ((f >> 1) & 0x7);
+    return static_cast<uint8_t>(((f >> 1) & 0x7));
 }
 
 uint32_t Lib__runAll(void) {
@@ -76,6 +76,6 @@ uint32_t Lib__runAll(void) {
 
 int main(void) {
     uint32_t result = Lib__runAll();
-    if (result != 122) return 1;
-    return 0;
+    if (result != 122) return 1U;
+    return 0U;
 }
